\name{OTUbase-package}
\alias{OTUbase-package}
\alias{OTUbase}
\docType{package}
\title{
The OTUbase package: A tool for organizing and manipulating Operational Taxonomic Unit data
}
\description{
 The \pkg{OTUbase} Base class for OTU data 
}
\details{
\tabular{ll}{
Package: \tab OTUbase\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2010-04-05\cr
License: \tab Artistic-2.0\cr
LazyLoad: \tab yes\cr
}
~~OTUbase includes a number of OTUset type classes which provide structure for OTU based data. These classes allow the user to store information that may be usefull in the analysis of OTUs. Slots are provided for sequence and quality values, OTU classifications, Sample identifications, and metadata associated with samples and OTUs. In addition, basic functions are provided for the analysis and visualization of the data. In addition to OTU type analysis, classification data is also supported with the TAXset classes.~~
}
\author{
 Daniel Beck - \email{danlbek@gmail.com}, Matt Settles - \email{msettles@uidaho.edu}, and James Foster 
Maintainer: Daniel Beck - \email{danlbek@gmail.com}
}
\references{
	An_introduction_to_OTUbase.pdf
}
\keyword{ package }

